+++ setup.cfg
@@ -102,6 +102,8 @@ filterwarnings =
     once::PendingDeprecationWarning
     ignore:Model '.*' was already registered. Reloading models is not advised.*:RuntimeWarning
 addopts = --doctest-modules
+# can be enabled to emit log message immediately
+log_cli = False
 
 [flake8]
 # A = builtins
