+++ tests/files/data/gebieden_ggwgebieden_bestaat_uit_buurten_delete.gobevents
@@ -1,5 +1,2 @@
-03630950000015.1.AMSBI.M50e|{"event_type": "ADD", "event_id": 13455962, "source_id": "03630950000015.1.AMSBI.M50e", "last_event": null, "catalog": "rel", "collection": "gbd_ggw_gbd_brt_bestaat_uit_buurten", "source": "GOB"}|{"entity": {"id": "03630950000015.1.AMSBI.M50e", "dst_id": "03630023754008", "src_id": "03630950000015", "_source": "GOB", "_version": "0.1", "_source_id": "03630950000015.1.AMSBI.M50e", "bronwaarde": "M50e", "derivation": "code", "dst_source": "AMSBI", "src_source": "AMSBI", "_application": "Basisinformatie", "dst_volgnummer": 1, "src_volgnummer": 1, "_last_dst_event": 18317, "_last_src_event": 139551, "eind_geldigheid": "2019-10-03T00:00:00.000000", "_expiration_date": null, "begin_geldigheid": "2015-01-01T00:00:00.000000"}, "_source_id": "03630950000015.1.AMSBI.M50e", "_entity_source_id": "03630950000015.1.AMSBI.M50e"}
-03630950000015.2.AMSBI.M50e|{"event_type": "ADD", "event_id": 13456183, "source_id": "03630950000015.2.AMSBI.M50e", "last_event": null, "catalog": "rel", "collection": "gbd_ggw_gbd_brt_bestaat_uit_buurten", "source": "GOB"}|{"entity": {"id": "03630950000015.2.AMSBI.M50e", "dst_id": "03630023754008", "src_id": "03630950000015", "_source": "GOB", "_version": "0.1", "_source_id": "03630950000015.2.AMSBI.M50e", "bronwaarde": "M50e", "derivation": "code", "dst_source": "AMSBI", "src_source": "AMSBI", "_application": "Basisinformatie", "dst_volgnummer": 1, "src_volgnummer": 2, "_last_dst_event": 18317, "_last_src_event": 139551, "eind_geldigheid": null, "_expiration_date": null, "begin_geldigheid": "2015-01-01T00:00:00.000000"}, "_source_id": "03630950000015.2.AMSBI.M50e", "_entity_source_id": "03630950000015.2.AMSBI.M50e"}
-03630950000016.1.AMSBI.N65d|{"event_type": "ADD", "event_id": 13456184, "source_id": "03630950000016.1.AMSBI.N65d", "last_event": null, "catalog": "rel", "collection": "gbd_ggw_gbd_brt_bestaat_uit_buurten", "source": "GOB"}|{"entity": {"id": "03630950000016.1.AMSBI.N65d", "dst_id": "03630000000658", "src_id": "03630950000016", "_source": "GOB", "_version": "0.1", "_source_id": "03630950000016.1.AMSBI.N65d", "bronwaarde": "N65d", "derivation": "code", "dst_source": "AMSBI", "src_source": "AMSBI", "_application": "Basisinformatie", "dst_volgnummer": 1, "src_volgnummer": 1, "_last_dst_event": 18317, "_last_src_event": 139551, "eind_geldigheid": "2019-10-03T00:00:00.000000", "_expiration_date": null, "begin_geldigheid": "2014-02-20T00:00:00.000000"}, "_source_id": "03630950000016.1.AMSBI.N65d", "_entity_source_id": "03630950000016.1.AMSBI.N65d"}
-03630950000015.1.AMSBI.M50e|{"event_type": "DELETE", "event_id": 89300409, "source_id": "03630950000015.1.AMSBI.M50e", "last_event": 13455962, "catalog": "rel", "collection": "gbd_ggw_gbd_brt_bestaat_uit_buurten", "source": "GOB"}|{"_source_id": "03630950000015.1.AMSBI.M50e", "_entity_source_id": "03630950000015.1.AMSBI.M50e"}
-
+gebieden.ggwgebieden.03630950000000|{"event_type": "ADD", "dataset_id": "gebieden", "table_id": "ggwgebieden"}|{"code": "DX01", "naam": "Centrum-West", "geometrie": "POLYGON ((121734.944 487124.933, 121570.358 486848.056, 121561.071 486825.241, 121530.353 486772.016, 121734.944 487124.933))", "volgnummer": 1, "documentdatum": "2017-10-10", "identificatie": "03630950000000", "documentnummer": "gmb-2018-6914.pdf", "eind_geldigheid": "2019-10-03", "begin_geldigheid": "2014-02-20", "registratiedatum": "2019-10-22T12:43:04.568460", "ligt_in_stadsdeel": null, "bestaat_uit_buurten": [{"identificatie": "03630023754008", "volgnummer": 1, "begin_geldigheid": "2006-06-12T00:00:00.000000", "eind_geldigheid": null}]}
+gebieden.ggwgebieden.03630950000000|{"event_type": "MODIFY", "dataset_id": "gebieden", "table_id": "ggwgebieden"}|{"code": "DX01", "naam": "Centrum-West", "geometrie": "POLYGON ((121734.944 487124.933, 121570.358 486848.056, 121561.071 486825.241, 121530.353 486772.016, 121734.944 487124.933))", "volgnummer": 1, "documentdatum": "2017-10-10", "identificatie": "03630950000000", "documentnummer": "gmb-2018-6914.pdf", "eind_geldigheid": "2019-10-03", "begin_geldigheid": "2014-02-20", "registratiedatum": "2019-10-22T12:43:04.568460", "ligt_in_stadsdeel": null, "bestaat_uit_buurten": null}
